Solution to Find Greater Number

+++ Find Greater Number@@6788dd63490637cfc44e1b8e/code.c
@@ -1,9 +1,8 @@
 #include <stdio.h>
-#include <math.h>
 int main(){
     int a, b;
     scanf("%d %d", &a, &b);
-    int m = (int)fmax(a, b);
+    int m = fmax(a, b);
     printf("%d", m);
     return 0;
 }
